tach Controller
