+++ .gitignore
@@ -16,3 +16,4 @@ yarn-error.log
 /public/js
 /public/mix-manifest.json
 /public/ckeditor
+/coverage
